Fix sed regex hyphen placement
`sed` complained about the former rule contained an illegal charanger
range. According to [0], 9.3.5 RE Bracked Expression (7) one's supposed
to have the - be either first or last for it to be treated literally.
This is also different from some languages, e.g. python, in which the
previous regex works just fine.

[0]: https://www2.cs.sfu.ca/~anoop/courses/CMPT-379-Fall-2004/posix-regexp.html